+++ mpu/stock_io.py
@@ -5,22 +5,8 @@ from pathlib import Path
 
 import pandas as pd
 
-from mpu.pyopenxl_utils import format_and_save_df, EXCEL_ENGINE
-
-COLUMNS_FORMAT = {
-    "idArticle": {"hidden": True},
-    "Local Name": {"hidden": True},
-    "English Name": {"width": 7.5},
-    "Exp. Name": {"hidden": True},
-    "Signed?": {"hidden": True},
-    "Playset?": {"hidden": True},
-    "Altered?": {"hidden": True},
-    "idCurrency": {"hidden": True},
-    "Currency Code": {"hidden": True},
-    "Price": {"color": "949494E8"},
-    "SuggestedPrice": {"color": "949494E8"},
-    "PriceApproval": {"color": "FFF0F8FF"},
-}
+from mpu.excel_formats import STOCK_COLUMNS_FORMAT
+from mpu.utils.pyopenxl_utils import format_and_save_df, EXCEL_ENGINE
 
 
 def get_stock_file_path(folder_path: Path) -> Path:
@@ -40,4 +26,4 @@ def save_stock_df_as_excel_formatted_file(df: pd.DataFrame, file_path: Path) -> 
     writer = pd.ExcelWriter(path=str(file_path), engine=EXCEL_ENGINE)
     df.to_excel(excel_writer=writer, index=False, engine=EXCEL_ENGINE)
 
-    format_and_save_df(df=df, writer=writer, format_config=COLUMNS_FORMAT)
+    format_and_save_df(df=df, writer=writer, format_config=STOCK_COLUMNS_FORMAT)