+++ mpu/utils/strategies_utils.py
@@ -1,10 +1,9 @@
 import enum
-import json
 import sys
 from pathlib import Path
-from typing import Any, Dict, NamedTuple, Optional
+from typing import Any, Dict, NamedTuple
 
-sys.path.append(str(Path(__file__).parent / "MPUStrategies"))
+sys.path.append(str(Path(__file__).parent.parent / "MPUStrategies"))
 
 from mpu_strategies.compute_current_price import CurrentPriceComputer
 from mpu_strategies.errors import SuitableExamplesShortage
@@ -16,20 +15,17 @@ class StrategiesOptions(NamedTuple):
     price_update: Dict[str, Any]
 
 
-def get_strategies_options(
-    strategies_options_path: Optional[Path],
-) -> StrategiesOptions:
+def get_strategies_options(config: dict) -> StrategiesOptions:
     """Returns the current strategy options"""
 
-    if strategies_options_path is not None:
-        with strategies_options_path.open("r") as strategies_options_file:
-            strategies_options = json.load(fp=strategies_options_file)
+    strategies_options = config.get("strategies_options", None)
+    if strategies_options is None or not strategies_options:
+        return StrategiesOptions(current_price={}, price_update={})
 
-            return StrategiesOptions(
-                current_price=strategies_options["current_price"],
-                price_update=strategies_options["price_update"],
-            )
-    return StrategiesOptions(current_price={}, price_update={})
+    return StrategiesOptions(
+        current_price=strategies_options["current_price"],
+        price_update=strategies_options["price_update"],
+    )
 
 
 CurrentPriceStrat = enum.Enum(