+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.5.0]
+- A lot more information on the stats
+- Not requesting the signed and altered cards
+- Switch to a mandatory config path for strategies' options
+- Adding options to change the number of cards requested per language, the default max results 
+and the requests split
 
 # [0.4.2] - 2021-01-11
 - Wider date column in stats output